Added Settings button, rework of notes, added alert, UI fixes
